pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-tps"
description = "Total Participation Statistics for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "killboard",
    "pvp",
    "statistics",
    "zkillboard",
]
license = { file = "LICENSE" }
authors = [
    { name = "BioBrute", email = "broodlk@gmail.com" },
]
requires-python = ">=3.10,<3.14"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.3.1,<5",
    "django-eveuniverse",
]
urls.Changelog = "https://github.com/BroodLK/aa-tps/blob/master/CHANGELOG.md"
urls.Tracker = "https://github.com/BroodLK/aa-tps/issues"

[tool.hatch.version]
path = "aatps/__init__.py"

[tool.hatch.build]
include = [
    "/aatps",
]

[tool.black]
line-length = 120
target-version = [ "py310" ]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = [ "aatps" ]
known_django = [ "django" ]
sections = [ "FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER" ]
